<Fix> 식단 관련 컴포넌트 로직 분리 완료
