+++ src/hooks/menu/useMenuData.ts
@@ -1,22 +1,22 @@
-//Daily는 ‘하루 3끼 + 스와이프’, Weekly는 ‘한 주 테이블’—요구사항과 레이아웃이 완전히 달라요. 한 컴포넌트에 섞으면 if 블록이 폭증하고 유지가 어려워집니다.
-//단일 책임을 위해서 커스텀 훅으로 분리합니다.
-//상태, 파생데이터, 핸들러 함수 및 유틸함수들을 분리해놓겠습니다.
 import { useMemo } from 'react';
 import { useQuery } from '@tanstack/react-query';
-import { getMenus } from '@/api/menu';
-import type { MenuItem } from '@/api/menu';
+import { getMenus, type MenuItem } from '@/api/menu';
+
+const ORDER: Array<MenuItem['menuCategory']> = ['BREAKFAST', 'LUNCH', 'DINNER'];
 
 export function useMenuData() {
   const {
     data = [],
     isLoading,
     error,
+    refetch,
   } = useQuery<MenuItem[]>({
     queryKey: ['menus'],
     queryFn: getMenus,
     staleTime: 5 * 60 * 1000,
   });
 
+  // 날짜별 그룹 + 끼니 순 정렬 + 날짜 정렬
   const groupedByDate = useMemo(() => {
     const map = new Map<string, MenuItem[]>();
     for (const m of data) {
@@ -24,24 +24,22 @@ export function useMenuData() {
       map.get(m.date)!.push(m);
     }
     for (const [, arr] of map) {
-      arr.sort(
-        (a, b) =>
-          ['BREAKFAST', 'LUNCH', 'DINNER'].indexOf(a.menuCategory) -
-          ['BREAKFAST', 'LUNCH', 'DINNER'].indexOf(b.menuCategory),
-      );
+      arr.sort((a, b) => ORDER.indexOf(a.menuCategory) - ORDER.indexOf(b.menuCategory));
     }
     return Array.from(map.entries()).sort(([a], [b]) => a.localeCompare(b));
   }, [data]);
 
-  const todayKey = (() => {
+  // 키 모음 / 오늘 키 / 헬퍼
+  const keys = groupedByDate.map(([k]) => k);
+  const todayKey = useMemo(() => {
     const now = new Date();
     const mm = String(now.getMonth() + 1).padStart(2, '0');
     const dd = String(now.getDate()).padStart(2, '0');
     const dow = ['일', '월', '화', '수', '목', '금', '토'][now.getDay()];
     return `${mm}.${dd} (${dow})`;
-  })();
-  const getByDate = (key: string) => groupedByDate.find(([d]) => d === key)?.[1] ?? [];
-  const keys = groupedByDate.map(([k]) => k);
+  }, []);
 
-  return { isLoading, error, groupedByDate, keys, todayKey, getByDate };
+  const getByDate = (key: string) => groupedByDate.find(([d]) => d === key)?.[1] ?? [];
+
+  return { isLoading, error, refetch, groupedByDate, keys, todayKey, getByDate };
 }
